refactor(frontend): use configs index import
