+++ frontend/__tests__/.server/domain/mappers/marital-status.dto.mapper.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { ServerConfig } from '~/.server/configs/server.config';
+import type { ServerConfig } from '~/.server/configs';
 import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusEntity } from '~/.server/domain/entities';
 import { MaritalStatusDtoMapperImpl } from '~/.server/domain/mappers';
