+++ frontend/__tests__/.server/domain/mappers/preferred-communication-method.dto.mapper.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { ServerConfig } from '~/.server/configs/server.config';
+import type { ServerConfig } from '~/.server/configs';
 import type { PreferredCommunicationMethodDto } from '~/.server/domain/dtos';
 import type { PreferredCommunicationMethodEntity } from '~/.server/domain/entities';
 import { PreferredCommunicationMethodDtoMapperImpl } from '~/.server/domain/mappers';
