+++ frontend/app/.server/domain/mappers/marital-status.dto.mapper.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from 'inversify';
 
-import type { ServerConfig } from '~/.server/configs/server.config';
+import type { ServerConfig } from '~/.server/configs';
 import { SERVICE_IDENTIFIER } from '~/.server/constants/service-identifier.contant';
 import type { MaritalStatusDto } from '~/.server/domain/dtos';
 import type { MaritalStatusEntity } from '~/.server/domain/entities';
